LayoutTests/media/modern-media-controls/media-controller/media-controller-compact.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 240px;" controls autoplay></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>MediaController</code> sizing behavior with compact mode.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

shouldBecomeEqual("!!(mediaController.controls.layoutTraits & LayoutTraits.Compact)", "true", finishJSTest);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
